fix: The style error problem of calInfo
